+++ src/LevelSetEvolution/Evolver/FiniteDifferenceEvolver.jl
@@ -67,6 +67,7 @@ end
 
 get_min_dof_spacing(m::FiniteDifferenceEvolver) = minimum(m.params.Δ)
 get_ls_space(m::FiniteDifferenceEvolver) = m.space
+get_max_steps(m::FiniteDifferenceEvolver) = m.params.max_steps
 
 # Compute the time step for the `FiniteDifferenceEvolver`.
 function compute_Δt(::FiniteDifferenceEvolver,Δ,γ,φ,vel)